% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeGrader-package.R
\docType{package}
\name{codeGrader-package}
\alias{codeGrader-package}
\title{codeGrader: Automated Execution Checks for Student R Homework}
\description{
Start with \code{\link[=grader_check_setup]{grader_check_setup()}}, then run \code{\link[=codeGrader]{codeGrader()}}.
See the package README for the recommended first-run sequence and privacy notes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mcvanderbilt/code-library}
  \item Report bugs at \url{https://github.com/mcvanderbilt/code-library/issues}
}

}
\author{
\strong{Maintainer}: Matthew C. Vanderbilt

Authors:
\itemize{
  \item Matthew C. Vanderbilt
}

}
\keyword{internal}
